test: extract Figma Plugin API mocking functions and split test cases for maintainability
